Updated runner code to use the new writer implicits.
Removed ability to open properties from file but that can be added later.
